ai-system/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
telnyx>=2.0.0
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.33.0
accelerate>=0.23.0
scipy
ftfy
openai-whisper
soundfile
requests
python-multipart
safetensors
opencv-python
